Make gpoly's type 16bit
git-svn-id: svn://svn.code.sf.net/p/libgarmin/svn@304 1ee56e69-fe3a-0410-901e-914b4450df7a

+++ libgarmin/dev/src/libgarmin_priv.h
@@ -126,7 +126,7 @@ struct gpoint {
 struct gpoly {
 	struct gar_subdiv *subdiv;
 	u_int16_t n;
-	u_int8_t type;
+	u_int16_t type;
 	u_int32_t lbloffset;
 	struct gcoord c;
 	unsigned dir:1,
